refactor workspace state cell to its own component
